refactor: use Grid instead of Container & named exports
