+++ src/app/Routes.tsx
@@ -1,14 +1,14 @@
-import React, { Suspense, lazy, useEffect } from 'react';
+import React, { Suspense } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { PageLoader } from 'app/modules/common/PageLoader';
 
-import Auth from 'app/modules/auth';
+import SignIn from 'app/modules/SignIn';
 
 function Routes() {
   return (
     <Suspense fallback={<PageLoader />}>
       <Switch>
-        <Route exact path="/" render={() => <Auth />} />
+        <Route exact path="/" render={() => <SignIn />} />
       </Switch>
     </Suspense>
   );
