+++ src/app/components/inputs/buttons/ContainedButton/story.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import Comp from '.';
+import { ContainedButton } from '.';
 import Providers from 'app/Providers';
 
 storiesOf('Inputs|Buttons', module).add('Contained - Enabled', () => (
   <>
     <Providers>
-      <Comp text="Button" />
+      <ContainedButton text="Button" />
     </Providers>
   </>
 ));
@@ -15,7 +15,7 @@ storiesOf('Inputs|Buttons', module).add('Contained - Enabled', () => (
 storiesOf('Inputs|Buttons', module).add('Contained - Disabled', () => (
   <>
     <Providers>
-      <Comp text="Button" disabled />
+      <ContainedButton text="Button" disabled />
     </Providers>
   </>
 ));
