+++ src/app/components/inputs/textfields/SingleLineTextField/story.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import Comp from '.';
+import { SingleLineTextField } from '.';
 import Providers from 'app/Providers';
 
 storiesOf('Inputs|TextFields', module).add('Login Username', () => (
   <div style={{ background: '#fff' }}>
     <Providers>
-      <Comp label="Username" id="login-username" />
+      <SingleLineTextField value="Username" id="login-username" />
     </Providers>
   </div>
 ));
