+++ src/index.css
@@ -56,9 +56,16 @@
 
 body {
   margin: 0;
+  width: 100vw;
+  height: 100vh;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   background-color: #ebebeb;
+}
+
+#root {
+  width: 100%;
+  height: 100%;
 }
 
 code {
